move multiclass likelihood tests into separate test file
